refactor: Remove components folder
