+++ web/src/common/DynamicTable.tsx
