+++ web/src/feature/calculation_input/CalculationInput.tsx
@@ -3,12 +3,12 @@ import { Card } from '../../common/Card'
 import styled from 'styled-components'
 import { FluidDialog } from '../package_dialog/FluidDialog'
 import { useState } from 'react'
-import { MultiflashResponse } from '../../../api/generated'
+import { MultiflashResponse } from '../../api/generated'
 import { AxiosError, AxiosResponse } from 'axios'
-import MercuryAPI from '../../../api/MercuryAPI'
+import MercuryAPI from '../../api/MercuryAPI'
 import { FeedFlowInput } from './FeedFlowInput'
-import { TComponentProperty, TPackage, TResults } from '../../../types'
-import useLocalStorage from '../../../hooks/useLocalStorage'
+import { TComponentProperty, TPackage, TResults } from '../../types'
+import useLocalStorage from '../../hooks/useLocalStorage'
 import { TempOrPressureInput } from './TempOrPressureInput'
 
 const FlexContainer = styled.div`
