+++ web/src/feature/package_dialog/ComponentTable.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import { EdsProvider, Input, Table } from '@equinor/eds-core-react'
-import { TComponentRatios, TComponentProperty } from '../../../types'
+import { TComponentRatios, TComponentProperty } from '../../types'
 
 const ComponentTableContainer = styled.div`
   display: flex;
