+++ web/src/feature/results/MoleTable.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react'
 import { MoleTable } from './MoleTable'
 import '@testing-library/jest-dom/extend-expect'
-import { mockComponentProperties, mockResults } from '../../../setupTests'
+import { mockComponentProperties, mockResults } from '../../setupTests'
 
 test('renders without crashing and displaying correct values in table', async () => {
   render(
