+++ web/src/pages/Main.tsx
@@ -1,14 +1,14 @@
 import styled from 'styled-components'
 import { Divider, Typography } from '@equinor/eds-core-react'
-import { Header } from '../components/common/Header'
+import { Header } from '../common/Header'
 import { useEffect, useState } from 'react'
-import { MoleTable } from '../components/feature/results/MoleTable'
-import { CalculationInput } from '../components/feature/calculation_input/CalculationInput'
+import { MoleTable } from '../feature/results/MoleTable'
+import { CalculationInput } from '../feature/calculation_input/CalculationInput'
 import MercuryAPI from '../api/MercuryAPI'
 import { AxiosResponse } from 'axios'
 import { ComponentProperties, ComponentResponse } from '../api/generated'
-import { PhaseTable } from '../components/feature/results/PhaseTable'
-import { MercuryWarning } from '../components/feature/results/MercuryWarning'
+import { PhaseTable } from '../feature/results/PhaseTable'
+import { MercuryWarning } from '../feature/results/MercuryWarning'
 import { TComponentProperty, TResults } from '../types'
 import { orderedComponents } from '../constants'
